Cellular fix only: fix tick wrap in cellular timeout.
When the return value of uPortTickTimeMs() was changed to int32_t (from int64_t) in commit 30dd3b089a340ba9ebcef340962decc9548762c7 back in August 2022, a few bits were missed: in this case the stop time is stored as a 64-bit integer in the cell context and, if the tick happens to fall within the timeout window of INT32_MAX, the stop time would be set beyond the range of int32_t.  Even if it were wrapped around correctly, the form of the check in the static keep going callback function means that it would immediately time out.  The fix is to change the stop time value to use int32_t and fix the check used in the keep going callback to calculate a difference.

Our thanks to warasilapm for spotting this and suggesting the fix.

+++ common/device/src/u_device_shared_cell.h
@@ -46,7 +46,7 @@ extern "C" {
 typedef struct {
     int32_t uart;
     uAtClientHandle_t at;
-    int64_t stopTimeMs;
+    int32_t stopTimeMs;
     int32_t pinPwrOn;
 } uDeviceCellContext_t;
